Convert nc_nlen and cn_namelen to u_char.
NAME_MAX is 255, so u_char is sufficient.

input and ok pedro@

+++ sys/msdosfs/msdosfs_lookup.c
@@ -195,7 +195,7 @@ msdosfs_lookup(void *v)
 	
 #ifdef MSDOSFS_DEBUG
 	printf("msdosfs_lookup(): dos version of filename '%.11s', length %d\n",
-	    dosfilename, cnp->cn_namelen);
+	    dosfilename, (int)cnp->cn_namelen);
 #endif
 
 	/*
